update: Add new features and some refactor

+++ src/main/java/com/easyaccess/app/common/providers/JwtTokenProvider.java
@@ -30,7 +30,7 @@ public class JwtTokenProvider {
       .getPayload();
   }
 
-  public String generateToken(String email, String role, Long userId) {
+  public String generateToken(String email, String role, int userId) {
     return Jwts.builder()
       .subject(email)
       .issuedAt(new Date())
